# apps/governance-ui/openapi/control-plane.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dal-obscura control-plane API",
    "summary": "Configuration, catalog, asset, policy, and session API for dal-obscura.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/plugins": {
      "get": {
        "summary": "List Plugins",
        "operationId": "list_plugins_v1_plugins_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "__Host-dal_obscura_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Dal Obscura Session"
            }
          },
          {
            "name": "__Host-dal_obscura_csrf",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Dal Obscura Csrf"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginListResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Precondition Required"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/v1/plugins/{kind}/{plugin_id}/lifecycle": {
      "patch": {
        "summary": "Set Plugin Lifecycle",
        "operationId": "set_plugin_lifecycle_v1_plugins__kind___plugin_id__lifecycle_patch",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "catalog",
                "table_format"
              ],
              "type": "string",
              "title": "Kind"
            }
          },
          {
            "name": "plugin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plugin Id"
            }
          },
          {
            "name": "__Host-dal_obscura_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Dal Obscura Session"
            }
          },
          {
            "name": "__Host-dal_obscura_csrf",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Dal Obscura Csrf"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PluginLifecycleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginLifecycleResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Precondition Required"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "current_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Revision"
          },
          "field_errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ApiFieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Errors"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ApiError",
        "description": "Machine-readable error metadata shared by all HTTP rejections."
      },
      "ApiErrorResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Detail"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "error"
        ],
        "title": "ApiErrorResponse",
        "description": "Public HTTP error envelope documented for generated clients."
      },
      "ApiFieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message",
          "type"
        ],
        "title": "ApiFieldError",
        "description": "Validation evidence without submitted values or secrets."
      },
...
      "PluginDescriptorResponse": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "catalog",
              "table_format"
            ],
            "title": "Kind"
          },
          "plugin_id": {
            "type": "string",
            "title": "Plugin Id"
          },
          "api_version": {
            "type": "string",
            "title": "Api Version"
          },
          "config_version": {
            "type": "integer",
            "title": "Config Version"
          },
          "distribution": {
            "type": "string",
            "title": "Distribution"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          },
          "output_formats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Output Formats"
          },
          "handle_versions": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Handle Versions"
          },
          "config_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config Schema"
          },
          "status": {
            "type": "string",
            "enum": [
              "admitted",
              "incompatible"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "plugin_id",
          "api_version",
          "config_version",
          "distribution",
          "version",
          "display_name",
          "capabilities",
          "output_formats",
          "handle_versions",
          "config_schema",
          "status"
        ],
        "title": "PluginDescriptorResponse",
        "description": "Admitted plugin descriptor exposed to operator tooling."
      },
      "PluginLifecycleRequest": {
        "properties": {
          "target": {
            "type": "string",
            "enum": [
              "enabled",
              "draining",
              "disabled",
              "revoked",
              "removed"
            ],
            "title": "Target"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target"
        ],
        "title": "PluginLifecycleRequest",
        "description": "Requested process-local admission lifecycle transition."
      },
      "PluginLifecycleResponse": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "catalog",
              "table_format"
            ],
            "title": "Kind"
          },
          "plugin_id": {
            "type": "string",
            "title": "Plugin Id"
          },
          "lifecycle": {
            "type": "string",
            "enum": [
              "enabled",
              "draining",
              "disabled",
              "revoked",
              "removed"
            ],
            "title": "Lifecycle"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "plugin_id",
          "lifecycle"
        ],
        "title": "PluginLifecycleResponse",
        "description": "Result of one explicit plugin lifecycle transition."
      },
      "PluginListResponse": {
        "properties": {
          "plugins": {
            "items": {
              "$ref": "#/components/schemas/PluginDescriptorResponse"
            },
            "type": "array",
            "title": "Plugins"
          },
          "pairs": {
            "items": {
              "$ref": "#/components/schemas/PluginPairResponse"
            },
            "type": "array",
            "title": "Pairs"
          },
          "states": {
            "items": {
              "$ref": "#/components/schemas/PluginStateResponse"
            },
            "type": "array",
            "title": "States"
          }
        },
        "type": "object",
        "required": [
          "plugins",
          "pairs",
          "states"
        ],
        "title": "PluginListResponse",
        "description": "Bounded admitted plugin registry response."
      },
      "PluginPairResponse": {
        "properties": {
          "catalog_plugin_id": {
            "type": "string",
            "title": "Catalog Plugin Id"
          },
          "format_plugin_id": {
            "type": "string",
            "title": "Format Plugin Id"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          },
          "handle_versions": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Handle Versions"
          },
          "status": {
            "type": "string",
            "enum": [
              "admitted",
              "incompatible"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "catalog_plugin_id",
          "format_plugin_id",
          "capabilities",
          "handle_versions",
          "status"
        ],
        "title": "PluginPairResponse",
        "description": "One catalog/table-format compatibility result."
      },
      "PluginStateResponse": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "catalog",
              "table_format"
            ],
            "title": "Kind"
          },
          "plugin_id": {
            "type": "string",
            "title": "Plugin Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "enabled",
              "not_installed",
              "incompatible"
            ],
            "title": "Status"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "lifecycle": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "enabled",
                  "draining",
                  "disabled",
                  "revoked",
                  "removed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Lifecycle"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "plugin_id",
          "status"
        ],
        "title": "PluginStateResponse",
        "description": "Allowlisted plugin lifecycle status without factory imports."
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}